Consolidate custom-named prompt spec files into standard spec files
Integrate tab navigation and namespace display tests from custom-named
spec files into their corresponding standard spec files.

Changes:
- Merge promptTable-tabs.spec.tsx tests into promptTable.spec.tsx
- Merge promptDrawer-namespace.spec.tsx tests into promptDrawer.spec.tsx
- Delete custom-named spec files after integration

This consolidation ensures each component has a single comprehensive
spec file containing all its tests, rather than tests split across
multiple files with custom suffixes.
